==> 5-annotation/3-interpret/2_summarize_cleaned_vep_output.ipynb <==
# %%

from all_funx import *

# %%

# import
path  = "/Users/mariapalafox/Desktop/BOLDcaution/WGS/5-annotation/3-summarize/2-summarize-v4/"

indel12 = pd.read_csv('1_clean_vep_output-indel-cad12-v2.csv')
indel63 = pd.read_csv('1_clean_vep_output-indel-cad63-v2.csv')

snp12 = pd.read_csv('1_clean_vep_output-snp-cad12-v2.csv')
snp63 = pd.read_csv('1_clean_vep_output-snp-cad63-v2.csv')

dfs = [indel12, indel63, snp12, snp63]
dfnames = ["indel12", "indel63", "snp12", "snp63"]

snps = [snp12, snp63]
snpnames = ["snp12", "snp63"]

# save_describe(indel12, path + 'indel12')
# save_describe(snp12, path + 'snp12')
# save_describe(indel63, path + 'indel63')
# save_describe(snp63, path + 'snp63')

# %%

# QC total genes in dfs made with fillter given_ref_match == False
indel12_false = indel12[indel12['GIVEN_REF_MATCH'] == False]
snp12_false = snp12[snp12['GIVEN_REF_MATCH'] == False]
indel63_false = indel63[indel63['GIVEN_REF_MATCH'] == False]
snp63_false = snp63[snp63['GIVEN_REF_MATCH'] == False]
dfs_false = [indel12_false, indel63_false, snp12_false, snp63_false]
dfnames_false = ["indel12_false", "indel63_false", "snp12_false", "snp63_false"]

# %%

def save_colValueCount(colname, dfs, dfnames, PATH):
    #  funxtion to save value counts of a column in a multi-sheet excel
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + colname + '-checkValueCount' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            colval = checkColumnValue(df, colname)
            # save excel sheet
            colval.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

def save_dtypes(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'dtypes' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            typedf = df.dtypes
            # save excel sheet
            typedf.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

def save_dtypes(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'dtypes' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            typedf = df.dtypes
            # save excel sheet
            typedf.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

save_dtypes(dfs, dfnames, path)

# %%

# output file path
path  = "/Users/mariapalafox/Desktop/BOLDcaution/WGS/5-annotation/3-interpret/2-summarize-v4/"

# %%

save_dtypes(dfs, dfnames, path)

# %%

def save_describeNumeric(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'describeNumeric' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            desdf = i.describe(include='number')
            describee = pd.DataFrame(desdf).T
            # save excel sheet
            describee.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

def save_describeNumeric(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'describeNumeric' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            desdf = i.describe(include='number')
            describee = pd.DataFrame(desdf).T
            # save excel sheet
            describee.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

# describeNumberic
save_describeNumeric(dfs, dfnames, path)

# %%

def save_describeNumeric(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'describeNumeric' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            desdf = df.describe(include='number')
            describee = pd.DataFrame(desdf).T
            # save excel sheet
            describee.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

# describeNumberic
save_describeNumeric(dfs, dfnames, path)

# %%

def save_describeNull(dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'describeNull' + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        nan_column_from_1 =['cDNA_position', 'CDS_position', 'Protein_position', 'DISTANCE', 'EXONnumber', 'INTRONnumber', 'gnomADg_AF', 'gnomADg_AFR_AF', 'gnomADg_AMI_AF', 'gnomADg_AMR_AF', 'gnomADg_ASJ_AF', 'gnomADg_EAS_AF', 'gnomADg_FIN_AF', 'gnomADg_MID_AF', 'gnomADg_NFE_AF', 'gnomADg_OTH_AF', 'gnomADg_SAS_AF','MAX_AF', 'CLIN_SIG']
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            df_with_nan = df[nan_column_from_1].copy()
            sumdf = df_with_nan.isna().sum()
            # save excel sheet
            sumdf.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

# describeNull
save_describeNull(dfs, dfnames, path)

# %%

def save_uniqueCount(col, dfs, dfnames, PATH):
    #  dfs: list of dataframes
    #  dfnames: list of names of dataframes
    #  colname: name of column to check value counts
    #  PATH: path to save excel
    output =  PATH + 'uniqueCount-' + col + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            allCol =  []
            uniqCol = []
            lengthAll = len(df[col])
            lengthUniq = len(df[col].unique())
            colname1 = 'total_' + col
            colname2 = 'unique_' + col
            allCol.append(colname1)
            uniqCol.append(colname2)
            allCol.append(lengthAll)
            uniqCol.append(lengthUniq)
            uniqueDF = pd.DataFrame([allCol, uniqCol])
            # save excel sheet
            uniqueDF.to_excel(writer, sheet_name=dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

# unique values
count_unique_value_columns = ['keyID38','Gene','SYMBOL', 'CCDS', 'Feature', 'ENSP','SWISSPROT']

for i in count_unique_value_columns:
    save_uniqueCount(i, dfs, dfnames, path)

# %%

# unique values
count_unique_value_columns = ['Gene','SYMBOL', 'CCDS', 'Feature', 'ENSP','SWISSPROT']
# keyID38 is only in snp dfs
for i in count_unique_value_columns:
    save_uniqueCount(i, dfs, dfnames, path)

# %%

# keyID38 is only in snp dfs
save_uniqueCount('keyID38', snps, snpnames, path)

# %%

def save_crossTabIt(col1, col2, dfs, dfnames, PATH):
    # funx to save crosstab of columns in a sheet of an excel
    # col1: column name
    # col2: column name
    # dfs: list of dataframes
    # dfnames: list of names of dataframes
    output =  PATH + col1 + '-' + col2 + '-crossTabIt' + '.xlsx'
    for x, df in enumerate(dfs):
        dfname = dfnames[x]
        print(dfname)
        # function call
        x1, s1 = save_crosstabit_sheet(df, col1, col2, True)
        with pd.ExcelWriter(output) as writer:
            x1.to_excel(writer, sheet_name = s1)
    print('done saving multi-sheet excel : ',  output)

# %%

# cross tab
# be consistent with order of sift and polyphen
# add feature_type
save_crossTabIt("CLIN_SIG", "ZYG", dfs, dfnames, path)
save_crossTabIt("IMPACT", "ZYG", dfs, dfnames, path)
save_crossTabIt("Consequence", "ZYG", dfs, dfnames, path)
save_crossTabIt("Consequence", "IMPACT", dfs, dfnames, path)
save_crossTabIt("SYMBOL", "IMPACT", dfs, dfnames, path)
save_crossTabIt("SYMBOL", "Consequence", dfs, dfnames, path)

# %%

def save_crossTabIt(col1, col2, dfs, dfnames, PATH):
    # funx to save crosstab of columns in a sheet of an excel
    # col1: column name
    # col2: column name
    # dfs: list of dataframes
    # dfnames: list of names of dataframes
    output =  PATH + 'crossTabIt-' + col1 + '-' + col2 + '.xlsx'
    for x, df in enumerate(dfs):
        dfname = dfnames[x]
        print(dfname)
        # function call
        x1, s1 = save_crosstabit_sheet(df, col1, col2, True)
        with pd.ExcelWriter(output) as writer:
            x1.to_excel(writer, sheet_name = s1)
    print('done saving multi-sheet excel : ',  output)

# %%

def save_crossTabIt(col1, col2, dfs, dfnames, PATH):
    # funx to save crosstab of columns in a sheet of an excel
    # col1: column name
    # col2: column name
    # dfs: list of dataframes
    # dfnames: list of names of dataframes
    output =  PATH + 'crossTabIt-' + col1 + '-' + col2 + '.xlsx'
    for x, df in enumerate(dfs):
        dfname = dfnames[x]
        print(dfname)
        # function call
        x1, s1 = save_crosstabit_sheet(df, col1, col2, True)
        with pd.ExcelWriter(output) as writer:
            x1.to_excel(writer, sheet_name = s1)
    print('done saving multi-sheet excel : ',  output)

# %%

# check Column Value calls
save_colValueCount('SYMBOL', dfs_false, dfnames_false, path) # QC
save_colValueCount('Feature_type', dfs, dfnames, path)
save_colValueCount('Consequence', dfs, dfnames, path)
save_colValueCount('ZYG', dfs, dfnames, path)
save_colValueCount('IMPACT', dfs, dfnames, path)
save_colValueCount('VARIANT_CLASS', dfs, dfnames, path)
save_colValueCount('SYMBOL', dfs, dfnames, path)
save_colValueCount('SYMBOL_SOURCE', dfs, dfnames, path)
save_colValueCount('BIOTYPE', dfs, dfnames, path)
save_colValueCount('GIVEN_REF_MATCH', dfs, dfnames, path)
save_colValueCount('CLIN_SIG', dfs, dfnames, path)
save_colValueCount('PHENO', dfs, dfnames, path)
save_colValueCount('SOMATIC', dfs, dfnames, path)

#### new check values #########
save_colValueCount('chr', dfs, dfnames, path)
save_colValueCount('Allele', dfs, dfnames, path)
save_colValueCount('REF_ALLELE', dfs, dfnames, path)
save_colValueCount('Amino_acids', dfs, dfnames, path)
save_colValueCount('Gene', dfs, dfnames, path)
save_colValueCount('STRAND', dfs, dfnames,path)
save_colValueCount('FLAGS', dfs, dfnames, path)
save_colValueCount('SWISSPROT', dfs, dfnames, path)
save_colValueCount('GENE_PHENO', dfs, dfnames, path)
save_colValueCount('MAX_AF_POPS', dfs, dfnames, path)
save_colValueCount('HIGH_INF_POS', dfs, dfnames, path)
save_colValueCount('MOTIF_NAME', dfs, dfnames, path)

# %%

####### snps only ########
save_colValueCount('SIFTprediction', snps, snpnames, path)
save_colValueCount('PolyPhenprediction', snps, snpnames, path)

# %%

# #snp only
save_crossTabIt("SIFTprediction", "ZYG", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "ZYG", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "Consequence", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "Consequence", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "SYMBOL", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "SYMBOL", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "CLIN_SIG", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "CLIN_SIG", snps, snpnames, path)

# %%

def save_crossTabIt(col1, col2, dfs, dfnames, PATH):
    # funx to save crosstab of columns in a sheet of an excel
    # col1: column name
    # col2: column name
    # dfs: list of dataframes
    # dfnames: list of names of dataframes
    output =  PATH + 'crossTabIt-' + col1 + '-' + col2 + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            x1, s1 = save_crosstabit_sheet(df, col1, col2, True)
            x1.to_excel(writer, sheet_name = s1)
    print('done saving multi-sheet excel : ',  output)

# %%

# cross tab
# be consistent with order of sift and polyphen
# add feature_type
save_crossTabIt("CLIN_SIG", "ZYG", dfs, dfnames, path)

# %%

def save_crossTabIt(col1, col2, dfs, dfnames, PATH):
    # funx to save crosstab of columns in a sheet of an excel
    # col1: column name
    # col2: column name
    # dfs: list of dataframes
    # dfnames: list of names of dataframes
    output =  PATH + 'crossTabIt-' + col1 + '-' + col2 + '.xlsx'
    with pd.ExcelWriter(output) as writer:
        for x, df in enumerate(dfs):
            dfname = dfnames[x]
            print(dfname)
            # function call
            x1, s1 = save_crosstabit_sheet(df, col1, col2, True)
            x1.to_excel(writer, sheet_name = dfname)
    print('done saving multi-sheet excel : ',  output)

# %%

# cross tab
# be consistent with order of sift and polyphen
# add feature_type
save_crossTabIt("CLIN_SIG", "ZYG", dfs, dfnames, path)

# %%

# cross tab
# be consistent with order of sift and polyphen
# add feature_type
save_crossTabIt("CLIN_SIG", "ZYG", dfs, dfnames, path)
save_crossTabIt("IMPACT", "ZYG", dfs, dfnames, path)
save_crossTabIt("Consequence", "ZYG", dfs, dfnames, path)
save_crossTabIt("Consequence", "IMPACT", dfs, dfnames, path)
save_crossTabIt("SYMBOL", "IMPACT", dfs, dfnames, path)
save_crossTabIt("SYMBOL", "Consequence", dfs, dfnames, path)

# #snp only
save_crossTabIt("SIFTprediction", "ZYG", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "ZYG", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "Consequence", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "Consequence", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "SYMBOL", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "SYMBOL", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "CLIN_SIG", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "CLIN_SIG", snps, snpnames, path)

# %%

save_crossTabIt("PolyPhenprediction", "SYMBOL", snps, snpnames, path)

# %%

save_crossTabIt("SIFTprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "Feature_type", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "IMPACT", snps, snpnames, path)
save_crossTabIt("SIFTprediction", "CLIN_SIG", snps, snpnames, path)
save_crossTabIt("PolyPhenprediction", "CLIN_SIG", snps, snpnames, path)
